runtime_probe: Correct the `reg` int size of EC i2c components
The `expect` closure in EC component manifest file contains a
`reg` field, which represents the offset of the target register.
This CL updates the corresponding value type in the code to align
with the actual offset value type --- `uint8_t`, with additional
check when parsing the manifest file to avoid unexpected behaviors.

BUG=b:432594654
TEST=FEATURES=test emerge-<BOARD> chromeos-base/runtime_probe

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6787471
Auto-Submit: Yong Hong <yhong@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>

+++ runtime_probe/utils/ec_component_manifest.h
@@ -27,7 +27,7 @@ struct EcComponentManifest {
   struct Component {
     struct I2c {
       struct Expect {
-        uint32_t reg;
+        uint8_t reg;
         std::vector<uint8_t> write_data;
         std::optional<std::vector<uint8_t>> mask;
         std::optional<std::vector<uint8_t>> value;
